requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.8.0
python-dotenv==0.10.3
SQLAlchemy==1.3.5
Werkzeug==0.15.4
